Update README.md
Titanic Dataset Analyzed and Visualized
The Titanic dataset on which i have used business intelligence tool (TABLEAU) to analyze the data and visualized as a dashboard. Did Passengers Get Their Money's Worth?. 
The survival rate according to the fare.
The attachments are the Titanic dataset and the dashboard  i have created in Tableau.
The first project using Tableau.